1、支持主图和成交量子图、指标子图的价格/成交量小数位控制 2、支持绘制屏幕内最高最低价 3、优化其他显示细节，主要是参考火币app来进行优化

+++ kchartlib/src/main/java/com/github/tifezh/kchartlib/chart/base/IChartDraw.java
@@ -41,7 +41,7 @@ public interface IChartDraw<T> {
      * @param point
      * @return
      */
-    float getMaxValue(T point);
+    double getMaxValue(T point);
 
     /**
      * 获取当前实体中最小的值
@@ -49,7 +49,7 @@ public interface IChartDraw<T> {
      * @param point
      * @return
      */
-    float getMinValue(T point);
+    double getMinValue(T point);
 
     /**
      * 获取value格式化器
